ユーザーが編集できるようにした

+++ app/controllers/questions_controller.rb
@@ -35,12 +35,16 @@ class QuestionsController < ApplicationController
     question = question_details[:question]
     target = Question.find_by(id: question_id)
     begin
-      target.update!(
-        title: question[:title],
-        content: question[:content],
-        anonymous: question[:anonymous]
-      )
-      puts "変更できました"
+      if target.user_id == current_user[:id] then
+        target.update!(
+          title: question[:title],
+          content: question[:content],
+          anonymous: question[:anonymous]
+        )
+        puts "変更できました"
+      else
+        puts "投稿者以外は内容に変更を加えることができません"
+      end
     rescue ActiveRecord::RecordInvalid=> exception
       puts exception
       puts "変更できませんでした"   
@@ -73,9 +77,13 @@ class QuestionsController < ApplicationController
     question_id = params[:id]
     target = Question.find_by(id: question_id)
     begin
-      target.destroy!
-      puts "削除に成功しました"
-      # 保存成功時の処理
+      if target.user_id == current_user[:id] then
+        target.destroy!
+        puts "削除に成功しました"
+        # 保存成功時の処理
+      else
+        puts "投稿者以外は質問を削除することができません"
+      end
     rescue ActiveRecord::RecordInvalid => e
       puts e
       puts "削除に失敗しました"
